backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==1.10.14
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
